1.1.1 init time Config 4 space
