+++ Core/Error.py
@@ -1,0 +1,6 @@
+#!/usr/bin/env python
+#-*- coding=utf8 -*-
+__author__ = 'taochengwei'
+__date__ = '2015.11.25'
+__version__ = '1.1.1'
+__doc__ = '''Error handler'''
